+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * fix footer
   * allow GPX Split Download with Waypoints
   * bookmark Waypoints
+  * nicer copy to clipboard popover with BS Toast
 
 ### 20. June 2025
   * Mark camp sites with group_only=yes and permanent_camping=only with red